% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{d.invlogit}
\alias{d.invlogit}
\title{The derivative of the inverse logit function}
\usage{
d.invlogit(y)
}
\arguments{
\item{y}{The log-odds, a numeric value}
}
\value{
The derivative of inverse logit probability at point y
}
\description{
The derivative of the inverse logit function
}
\keyword{internal}
