% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebiasProgCV.R
\name{DebiasProgCV}
\alias{DebiasProgCV}
\title{The proposed debiasing (primal) program with cross-validation.}
\arguments{
\item{X}{The input design n*d matrix.}

\item{x}{The current query point, which is a 1*d array.}

\item{alpha_hat}{The estimated intercept of high-dimensional logistic regression, which is a number.}

\item{theta_hat}{The Lasso pilot estimator of high-dimensional logistic regression, which as a 1*d array}

\item{gamma_lst}{A numeric vector with candidate values for the regularization
parameter "\eqn{\gamma/n}". (Default: gamma_lst=NULL. Then, gamma_lst contains
41 equally spacing value between 0.001 and max(abs(x)).)}

\item{cv_fold}{The number of folds for cross-validation on the dual program.
(Default: cv_fold=5.)}

\item{cv_rule}{The criteria/rules for selecting the final value of the regularization
parameter "\eqn{\gamma/n}" in the dual program. (Default: cv_rule="1se". The candidate
choices include "1se", "minfeas", and "mincv".)}

\item{robust}{A boolean variable indicating whether we should remove the maximum and minimum when
calculating the average CV loss: (Default=FALSE.)}
}
\value{
A list that contains three elements.
\item{w_obs}{The final estimated weights by our debiasing program.}
\item{ll_obs}{The final value of the solution to our debiasing dual program.}
\item{gamma_n_opt}{The final value of the tuning parameter "\eqn{\gamma/n}" selected by cross-validation.}
\item{dual_loss}{A table indicating the dual loss for each fold under different tuning parameter.}
}
\description{
This function implements our proposed debiasing program that selects the tuning parameter
"\eqn{\gamma/n}" by cross-validation and returns the final debiasing weights.
}
\author{
Wenjie Guan, \email{wg285@cornell.edu}
}
