% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDLR_cf.R
\name{HDLR_cf}
\alias{HDLR_cf}
\title{Inference for High-Dimensional Logistic Regression with cross-fitting}
\arguments{
\item{X}{The input design n*d matrix.}

\item{Y}{The outcome variable, which is a n-dimensional vector.}

\item{x}{The current query point, which is a vector.}

\item{n_gamma}{Number of choices for the regularization parameter \eqn{\gamma/n}.}

\item{nfolds}{Number of folds in cross validation.}

\item{cv_rule}{Cross validation rule, candidate choices are '1se', 'mincv' and 'minfeas'.}

\item{refitting}{A boolean variable which indicates whether to refit on the Lasso support. Default is TRUE.}

\item{intercept}{A boolean variable which indicates whether to debias the intercept. Default is FALSE.}

\item{level}{The confidence level of the confidence interval. Default is 0.95.}
}
\value{
A list that contains the following elements.
\item{m}{The debiasing estimator for the log-odds, which is the inner product of x and coefficient.}
\item{sd}{The standard deviation for m.}
\item{prob}{The case probability 1 / (1 + exp(-m)).}
\item{prob_upper}{The upper limit of the confidence interval for the case probability.}
\item{prob_lower}{The lower limit of the confidence interval for the case probability.}
\item{level}{The confidence level.}
\item{m_pilot}{The Lasso pilot estimate for the log-odds.}
}
\description{
Inference for High-Dimensional Logistic Regression with cross-fitting
}
\examples{
\donttest{
require(MASS)
require(dplyr)
d = 200
n = 180
Sigma = array(0, dim = c(d,d)) + diag(d)
rho = 0.1
## AR(1) process
rho2 = 0.5
for (i in 1:d){
  for (j in 1:d){
      Sigma[i,j] = rho2 ^ (abs(i-j))
  }
}

## True regression coefficient
s_beta = 5
theta_0 = rep(0, d)
theta_0[1:s_beta] = sqrt(5)

## Generate the design matrix and outcomes via a logistic regression model with intercept 0.2.
set.seed(123)
X_sim = mvrnorm(n, mu = rep(0, d), Sigma) / 5
Y_sim = rbinom(n,size=1,prob=plogis(X_sim \%*\% theta_0 + 0.2))

## Current query point
x = rep(0, d)
x[c(1, 2, 3, 7, 8)] = c(1, 1/2, 1/4, 1/2, 1/8) / 5

res_cf = HDLR_cf(X_sim, Y_sim, x, n_gamma=20, cv_rule='1se', refitting=FALSE, intercept=TRUE)
cat("The 95\% confidence interval yielded by cross-fitting is [",
    res_cf$prob_lower, ", ",
    res_cf$prob_upper, "].\n", sep = "")

cat("The true probability is", plogis(x \%*\% theta_0 + 0.2))
}

}
\author{
Wenjie Guan, \email{wg285@cornell.edu}
}
