% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{invlogit}
\alias{invlogit}
\title{Auxiliary functions}
\usage{
invlogit(y)
}
\arguments{
\item{y}{The log-odds, a numeric value}
}
\value{
The corresponding inverse logit probability at point y
}
\description{
The inverse logit function
}
\keyword{internal}
