% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{robust_mean}
\alias{robust_mean}
\title{robust mean function}
\usage{
robust_mean(v)
}
\arguments{
\item{v}{an array}
}
\value{
The mean of v excluding the maximum and minimum
}
\description{
robust mean function
}
\author{
Wenjie Guan, \email{wg285@cornell.edu}
}
