% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftThres.R
\name{SoftThres}
\alias{SoftThres}
\title{The soft-thresholding function}
\usage{
SoftThres(theta, lamb)
}
\arguments{
\item{theta}{The input numeric vector.}

\item{lamb}{The thresholding parameter.}
}
\value{
The resulting vector after soft-thresholding.
}
\description{
The soft-thresholding function
}
\examples{
a = c(1,2,4,6)
SoftThres(theta=a, lamb=3)


}
\keyword{(x-\lambda)_+}.}
\keyword{This}
\keyword{\eqn{S_{\lambda}(x)=sign(x)\cdot}
\keyword{function}
\keyword{implements}
\keyword{internal}
\keyword{operator}
\keyword{soft-threshold}
\keyword{the}
