% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R, R/d.invlogit.R
\name{d.invlogit}
\alias{d.invlogit}
\title{the derivative of the sigmoid (inverse logit) function}
\usage{
d.invlogit(y)

d.invlogit(y)
}
\arguments{
\item{y}{an array containing values.}
}
\value{
The derivative at y.

The derivative at y.
}
\description{
the derivative of the sigmoid (inverse logit) function

the derivative of the sigmoid (inverse logit) function
}
\author{
Wenjie Guan, \email{wg285@cornell.edu}
}
