% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DualObj.R
\name{DualObj}
\alias{DualObj}
\title{the dual objective function}
\usage{
DualObj(X, x, theta_hat, alpha_hat, intercept = T, ll_cur, gamma_n = 0.05)
}
\arguments{
\item{X}{The input design n*d matrix.}

\item{x}{The current query point, which is a 1*d array.}

\item{theta_hat}{The Lasso pilot estimator of high-dimensional logistic regression, which as a 1*d array}

\item{alpha_hat}{????WJ}

\item{intercept}{????WJ}

\item{ll_cur}{The current value of the dual solution vector.}

\item{gamma_n}{The regularization parameter "\eqn{\gamma/n}". (Default: gamma_n=0.1.)}
}
\value{
The dual objective
}
\description{
the dual objective function
}
\author{
Wenjie Guan, \email{wg285@cornell.edu}
}
